move prompts logic to fileslist
License: MIT
